requirements.txt
streamlit
tensorflow
numpy
pillow
scikit-learn
matplotlib
opencv-python
seaborn


